feat: Ajout de l'authentificateur SupaBase
[ app/page.tsx
_components/taskModal.tsx ]: modification des fonctions API pour prendre en compte le user id

fetch.ts: modification des appels API pour création de session, changement d'authentifaction pour le token

taskContext.tsx: prise en charge du user.id dans l'appel d'origine

auth/page.tsx: création du formulaire d'authentification et quelsques gestions d'erreurs

useAuth.ts: renvoi d'un message si le user existe déjà

+++ app/_api/fetch.ts
@@ -1,20 +1,37 @@
 import { Task } from "../_utils/types";
+import { supabase } from "@/lib/supabase";
 
 const SUPABASE_URL = process.env.NEXT_PUBLIC_SUPABASE_URL;
 const SUPABASE_ANON_KEY = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY;
 
-export const getTasksAPI = async (): Promise<any> => {
+export const getTasksAPI = async (userId: string): Promise<Task[]> => {
   try {
-    const response = await fetch(`${SUPABASE_URL}/rest/v1/tasks?select=*`, {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        apikey: SUPABASE_ANON_KEY!,
-        Authorization: `Bearer ${SUPABASE_ANON_KEY}`,
-      },
-    });
+    const {
+      data: { session },
+    } = await supabase.auth.getSession();
+
+    if (!session) {
+      throw new Error("No session found");
+    }
+
+    const response = await fetch(
+      `${SUPABASE_URL}/rest/v1/tasks?select=*&user_id=eq.${userId}`,
+      {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+          apikey: SUPABASE_ANON_KEY!,
+          Authorization: `Bearer ${session.access_token}`,
+        },
+      }
+    );
+
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
 
     const data = await response.json();
+    console.log(data);
     return data;
   } catch (error) {
     console.error("Fetch error:", error);
@@ -22,7 +39,10 @@ export const getTasksAPI = async (): Promise<any> => {
   }
 };
 
-export const postTasksAPI = async (payload: Task): Promise<any> => {
+export const postTasksAPI = async (
+  payload: Task,
+  userId: string
+): Promise<any> => {
   const processedPayload = {
     title: payload.title,
     description: payload.description,
@@ -30,14 +50,24 @@ export const postTasksAPI = async (payload: Task): Promise<any> => {
     deadline: payload.deadline,
     priority: payload.priority,
     completed: false,
+    user_id: userId,
   };
+
   try {
+    const {
+      data: { session },
+    } = await supabase.auth.getSession();
+
+    if (!session) {
+      throw new Error("No session found");
+    }
+
     const response = await fetch(`${SUPABASE_URL}/rest/v1/tasks`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
         apikey: SUPABASE_ANON_KEY!,
-        Authorization: `Bearer ${SUPABASE_ANON_KEY}`,
+        Authorization: `Bearer ${session.access_token}`,
       },
       body: JSON.stringify(processedPayload),
     });
@@ -47,7 +77,10 @@ export const postTasksAPI = async (payload: Task): Promise<any> => {
   }
 };
 
-export const updateTasksAPI = async (payload: Task): Promise<any> => {
+export const updateTasksAPI = async (
+  payload: Task,
+  userId: string
+): Promise<any> => {
   const processedPayload = {
     title: payload.title,
     description: payload.description,
@@ -55,8 +88,16 @@ export const updateTasksAPI = async (payload: Task): Promise<any> => {
     deadline: payload.deadline,
     priority: payload.priority,
     completed: payload.completed,
+    user_id: userId,
   };
   try {
+    const {
+      data: { session },
+    } = await supabase.auth.getSession();
+
+    if (!session) {
+      throw new Error("No session found");
+    }
     const response = await fetch(
       `${SUPABASE_URL}/rest/v1/tasks?id=eq.${payload.id}`,
       {
@@ -64,7 +105,7 @@ export const updateTasksAPI = async (payload: Task): Promise<any> => {
         headers: {
           "Content-Type": "application/json",
           apikey: SUPABASE_ANON_KEY!,
-          Authorization: `Bearer ${SUPABASE_ANON_KEY}`,
+          Authorization: `Bearer ${session.access_token}`,
         },
         body: JSON.stringify(processedPayload),
       }
@@ -75,32 +116,58 @@ export const updateTasksAPI = async (payload: Task): Promise<any> => {
   }
 };
 
-export const deleteTasksAPI = async (id: number): Promise<any> => {
+export const deleteTasksAPI = async (
+  id: number,
+  userId: string
+): Promise<any> => {
   try {
-    const response = await fetch(`${SUPABASE_URL}/rest/v1/tasks?id=eq.${id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-        apikey: SUPABASE_ANON_KEY!,
-        Authorization: `Bearer ${SUPABASE_ANON_KEY}`,
-      },
-    });
+    const {
+      data: { session },
+    } = await supabase.auth.getSession();
+
+    if (!session) {
+      throw new Error("No session found");
+    }
+    const response = await fetch(
+      `${SUPABASE_URL}/rest/v1/tasks?id=eq.${id}&user_id=eq.${userId}`,
+      {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          apikey: SUPABASE_ANON_KEY!,
+          Authorization: `Bearer ${session.access_token}`,
+        },
+      }
+    );
   } catch (error) {
     console.error("Fetch error:", error);
     return null;
   }
 };
 
-export const getTaskIdAPI = async (id: number): Promise<any> => {
+export const getTaskIdAPI = async (
+  id: number,
+  userId: string
+): Promise<any> => {
   try {
-    const response = await fetch(`${SUPABASE_URL}/rest/v1/tasks?id=eq.${id}`, {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        apikey: SUPABASE_ANON_KEY!,
-        Authorization: `Bearer ${SUPABASE_ANON_KEY}`,
-      },
-    });
+    const {
+      data: { session },
+    } = await supabase.auth.getSession();
+
+    if (!session) {
+      throw new Error("No session found");
+    }
+    const response = await fetch(
+      `${SUPABASE_URL}/rest/v1/tasks?id=eq.${id}&user_id=eq.${userId}`,
+      {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+          apikey: SUPABASE_ANON_KEY!,
+          Authorization: `Bearer ${session.access_token}`,
+        },
+      }
+    );
     const data = await response.json();
     return data;
   } catch (error) {
